2 precios contado y credito

+++ database/migrations/2018_01_16_202132_create_camion_table.php
@@ -22,7 +22,7 @@ class CreateCamionTable extends Migration
             $table->string('marca');
             $table->string('color');
             $table->integer('ano');
-            $table->integer('peso');
+            $table->decimal('peso', 10, 2);
             $table->string('telefono');
             $table->timestamps();
         });
